mine cloning bug fix
I tried to contain this by adding a counter to mine tiles generated through minesweeper and setting that as the maximum amount of mine objects that can be generated in the scene.

As far as I've tested it, it works. But will need to test it further to make sure.